% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{The Euclidean algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A numeric scalar}

\item{y}{A numeric scalar}
}
\value{
The Euclidean algorithm on x and y. For the case of two integers, the GCD is yielded.
}
\description{
The algorithm takes two numeric scalars and applies the Euclidean
algorithm to it. In the case these two numeric scalars are
integers the greatest common divisor is yielded.
}
\details{
A description with pseudocode can be found at the Wikipedia page here:
https://en.wikipedia.org/wiki/Euclidean_algorithm.
}
