% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twofamousalgorithms-package.R
\docType{package}
\name{twofamousalgorithms-package}
\alias{twofamousalgorithms}
\alias{twofamousalgorithms-package}
\title{twofamousalgorithms: Implementations of the Euclidean algorithm and Dijkstra's algorithm}
\description{
This package includes implementations of two famous algorithms: the Euclidean algorithm for finding the greatest common divisor of two integers and Dijkstra algorithm for finding the shortest path in a graph.
}
\author{
\strong{Maintainer}: Linn F. \email{linfr259@student.liu.se}

Authors:
\itemize{
  \item John Moeller \email{johmo@student.liu.se}
}

}
\keyword{internal}
