% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A test dataset for Dijkstra's algorithm implementation}
\format{
\subsection{\code{wiki-graph}}{

A data frame with 18 rows and 3 variables:
\describe{
\item{v1, v2}{Edges of the graph}
\item{w}{Weight of the edge}
...
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27salgorithm}
}
\usage{
wiki_graph
}
\description{
A test dataset for Dijkstra's algorithm implementation
}
\keyword{datasets}
